🔧 fix: 보유 주식 평균가 타입 수정 #53

+++ BE/src/userStock/user-stock.entity.ts
@@ -21,7 +21,7 @@ export class UserStock {
   @Column({ nullable: false })
   quantity: number;
 
-  @Column('decimal', { nullable: false, precision: 10, scale: 5 })
+  @Column('float', { nullable: false, scale: 5 })
   avg_price: number;
 
   @UpdateDateColumn()
